UPDATE 0.07 ALPHA
Fixed the issue where every attack would count as a hit

You can now start a game and a miss would hand the turn over to the bot
The bot attacks instantaneous so maybe add a delay and the bots logic needs to be upgraded as when the bot hit you it attacks somewhere completely different

messed up the placement logic as it allows for ships to be placed next/on each other.

Currently the bots ship are visible for debugging. Logcat is correctly displaying error such as out of bounds and existing ship location but than the code places the ships on top of each other anyway...